+++ lib/src/core/models/DeveloperOptions.dart
@@ -2,12 +2,16 @@ import 'dart:convert';
 
 class DeveloperOptionsModel {
   bool enabled;
+  Map<String, bool> options = {};
+  Map<String, String> strings = {};
 
-  DeveloperOptionsModel({this.enabled = false});
+  DeveloperOptionsModel({this.enabled = false, this.options = const {}, this.strings = const {}});
 
   factory DeveloperOptionsModel.fromJson(Map<String, dynamic?> json) {
     return DeveloperOptionsModel(
-      enabled: json["enabled"] ?? false
+      enabled: json["enabled"] ?? false,
+      options: jsonDecode(json["options"]) ?? {},
+      strings: jsonDecode(json["strings"]) ?? {},
     );
   }
 
@@ -18,6 +22,7 @@ class DeveloperOptionsModel {
   Map<String, dynamic> toJson() {
     return {
       'enabled': this.enabled,
+      'options': this.options,
     };
   }