+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - 14.3.2021
+
+* Fix wrong named class in developer_options.dart
+* You can now store String/bool options and retrieve them.
+* You can now store String/String options and retrieve them.
+
+
 ## [0.1.1] - 14.3.2021
 
-* Still MVP, but can be used to store if developer options is enabled or not
+* Still WIP, but can be used to store if developer options is enabled or not